backend/requirements.txt
Flask
flask-cors
psycopg2-binary

